requirements.txt
langchain
langchain-community
langchain-anthropic
langchain-huggingface
lancedb
streamlit
streamlit-js-eval
python-dotenv
pypdf
docx2txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sentence-transformers
pyarrow
ipykernel
unstructured
python-docx
lxml
watchdog